LayoutTests/fast/events/input-events-fired-when-typing-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
Fired `onbeforeinput`!
PASS event.__lookupGetter__('inputType') is defined.
PASS event.__lookupGetter__('data') is defined.
PASS event.__lookupGetter__('dataTransfer') is defined.
PASS event.getTargetRanges is defined.
PASS event.target.id is expectedTargetID
PASS event.bubbles is true
PASS event.cancelable is true
PASS event.composed is true
Fired `oninput`!
PASS event.__lookupGetter__('inputType') is defined.
PASS event.__lookupGetter__('data') is defined.
PASS event.__lookupGetter__('dataTransfer') is defined.
PASS event.getTargetRanges is defined.
PASS event.target.id is expectedTargetID
PASS event.bubbles is true
PASS event.cancelable is false
PASS event.composed is true
Fired `onbeforeinput`!
PASS event.__lookupGetter__('inputType') is defined.
PASS event.__lookupGetter__('data') is defined.
PASS event.__lookupGetter__('dataTransfer') is defined.
PASS event.getTargetRanges is defined.
PASS event.target.id is expectedTargetID
PASS event.bubbles is true
PASS event.cancelable is true
PASS event.composed is true
Fired `oninput`!
PASS event.__lookupGetter__('inputType') is defined.
PASS event.__lookupGetter__('data') is defined.
PASS event.__lookupGetter__('dataTransfer') is defined.
PASS event.getTargetRanges is defined.
PASS event.target.id is expectedTargetID
PASS event.bubbles is true
PASS event.cancelable is false
PASS event.composed is true
a
